+++ examples/take_picture_local/src/app.c
@@ -3,7 +3,8 @@
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
-#include <assert.h>
+
+#include <xcore/assert.h>
 
 #include "io_utils.h"
 #include "app.h"
@@ -13,7 +14,7 @@ void user_app()
   // Initialize camera api
   camera_init();
 
-  int8_t image_buffer[APP_IMAGE_HEIGHT_PIXELS][APP_IMAGE_WIDTH_PIXELS][APP_IMAGE_CHANNEL_COUNT];
+  int8_t image_buffer[H][W][CH];
   uint8_t * image_ptr = (uint8_t *) &image_buffer[0][0][0];
 
   // set the input image to 0
@@ -24,7 +25,7 @@ void user_app()
 
   // grab a frame
   printf("Requesting image...\n");
-  assert(camera_capture_image(image_buffer) == 0);
+  xassert((camera_capture_image(image_buffer) == 0) && "Could not capture an image");
   printf("Image captured...\n");
 
   // stop the threads and camera stream